requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
tensorboard>=2.15.0
scikit-learn>=1.3.0
scipy>=1.11.0
Pillow>=10.0.0
tqdm>=4.66.0
tabulate>=0.9.0
